[refactor]:  organizando endpoints dentro da pasta Modules
