+++ src/LojaDoValdir/Dockerfile
@@ -1,6 +1,5 @@
-# Etapa base do runtime
 FROM mcr.microsoft.com/dotnet/aspnet:8.0 AS base
-LABEL maintainder "Hisllaylla Kézia"
+LABEL maintainer="Hisllaylla Kézia"
 WORKDIR /app
 EXPOSE 8080
 
@@ -8,11 +7,14 @@ EXPOSE 8080
 FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
 WORKDIR /src
 
+# Copiar o arquivo de projeto e restaurar dependências
 COPY src/LojaDoValdir/LojaDoValdir.csproj LojaDoValdir/
 RUN dotnet restore "LojaDoValdir/LojaDoValdir.csproj"
 
+# Copiar o código restante
 COPY src .
 
+# Compilar a aplicação
 WORKDIR "/src/LojaDoValdir"
 RUN dotnet build "LojaDoValdir.csproj" -c Release -o /app/build
 
@@ -23,5 +25,12 @@ RUN dotnet publish "LojaDoValdir.csproj" -c Release -o /app/publish
 # Etapa final
 FROM mcr.microsoft.com/dotnet/aspnet:8.0 AS final
 WORKDIR /app
+
+# Copiar a aplicação publicada
 COPY --from=publish /app/publish .
+
+# Configuração para rodar na porta 8080
+ENV ASPNETCORE_URLS=http://0.0.0.0:8080
+
+# Entrypoint para rodar a aplicação
 ENTRYPOINT ["dotnet", "LojaDoValdir.dll"]
