+++ src/LojaDoValdir/Usings/Using.cs
@@ -1,14 +1,10 @@
-global using Microsoft.OpenApi.Models;
 global using Microsoft.AspNetCore.Builder;
-global using Microsoft.AspNetCore.Hosting;
 global using Microsoft.AspNetCore.Http;
-global using Microsoft.EntityFrameworkCore;
+global using Microsoft.AspNetCore.Routing;
 global using Microsoft.Extensions.DependencyInjection;
-global using Microsoft.Extensions.Configuration;
-global using Microsoft.Extensions.Hosting;
-global using NSwag.AspNetCore;
 global using System;
 global using System.Collections.Generic;
 global using System.Linq;
-global using System.Threading.Tasks;
 global using LojaDoValdir.Enums;
+global using LojaDoValdir.Models;
+global using LojaDoValdir.Modules;
